pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "custom-subagents"
version = "0.1.0"
description = "Generate consumer-native agent definitions for Claude Code and Codex from a shared source."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "tomli>=2.0.1; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
shared-agents = "shared_agents.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
